pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Version-Aware Documentation Server for MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pydantic>=2.6.1",
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "aiohttp>=3.11.18",
    "openai>=1.76.0",
    "python-dotenv>=1.0.1",
]
